feat: add example in string and number types
