+++ src/@types/swaggerDefinition.d.ts
@@ -1,0 +1,13 @@
+export type SwaggerDefinition = {
+  [key: string]: {
+    type: string
+    format?: string
+    properties?: SwaggerDefinition
+    example?: any
+    items?: {
+      type: string
+      properties?: SwaggerDefinition
+      example?: string
+    };
+  };
+}
